Update pk_enc_utils.h
Finish pk encryption implementation and unit test. Works well.

The security parameter is a little bit tricky. According to Regev's paper, the default parameter should provide \lambda = 128, but that's given a smaller field. Perhaps needs to check the leftover hash lemma for detailed parameters. However, this should be the statistical security parameter, so \lambda = 40 is a popular choice in that sense.

Anyway, this should definitely be good to push into TFHE library for asymmetric encryption.